pyproject.toml
[project]
name = "trafficgram-is-subscribed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.12.0",
    "uvicorn>=0.34.3",
]
